Day 5 part 2 (rust): Refactor to support breaking on errors
This is to set the stage for automatically launching the debugger when
errors like out of bounds access occur. Ideally we'd "just" catch the
panic, like so:

	match panic::catch_unwind(|| { machine.execute() }) {
		Ok(_) => (),
		Err(_) => debug(&mut machine)
	}

But it's not possible to borrow mutably across the catch_unwind
boundary:

	error[E0277]: the type `&mut Machine` may not be safely transferred across an unwind boundary
	   --> src/main.rs:13:8
	    |
	13  |     match panic::catch_unwind(|| { machine.execute() }) {
	    |           ^^^^^^^^^^^^^^^^^^^ `&mut Machine` may not be safely transferred across an unwind boundary

So we'll have to propagate results up all the way to the top instead of
allowing panics to happen. This moves things in that direction by running
all of the argument reads through a single point in do_current_instruction.

Also, something in this makes part 2 work correctly.