Day 12 part 2 (rust): refactoring
Modeling the universe as a vector of mutable moons that have axes felt
a bit off. Because we process an axis at a time rather than a moon at
a time, the universe was left in an inconsistent state at all times
except at the start and end. By using three vectors of moon components
(one per axis) instead, we avoid that problem and make it more obvious
what's going on.